refactor(workspaces): unify serviceBusApi + workspacesApi into single backendApi
Collapses the two contextBridge namespaces and two handler interfaces into a
single BackendApi covering all four do-request methods. Future domains can be
added by extending one interface and one bridge object instead of multiplying
window globals.

- api-handler.ts + workspaces-api-handler.ts → backend-api.ts (BackendApi)
- All *FrontendClient classes now take BackendApi (constructor field renamed
  to backendApi for clarity)
- Electron preload exposes window.backendApi exposing all four methods;
  workspacesApi and serviceBusApi namespaces are gone
- WebServiceBusApiHandler + LocalStorageWorkspacesApiHandler → single
  WebBackendApi class. management/serviceBusManagement/messages still POST to
  the web backend; workspaces requests are handled locally against
  localStorage (web variant has no server-side workspace registry yet)
- Angular bindings provide WebBackendApi once and reuse it across all four
  *FrontendClient providers
