+++ libs/api-clients/clients/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/management-frontend-client';
 export * from './lib/messages-frontend-client';
 export * from './lib/service-bus-management-frontend-client';
-export * from './lib/api-handler';
+export * from './lib/backend-api';
 export * from './lib/workspaces-frontend-client';
-export * from './lib/workspaces-api-handler';
