--- libs/api-clients/clients/src/lib/api-handler.ts
@@ -1,8 +1,0 @@
-export interface ApiHandler {
-  serviceBusManagementDoRequest(
-    requestType: string,
-    request: unknown,
-  ): Promise<unknown>;
-  managementDoRequest(requestType: string, request: unknown): Promise<unknown>;
-  messagesDoRequest(requestType: string, request: unknown): Promise<unknown>;
-}
