+++ libs/api-clients/clients/src/lib/service-bus-management-frontend-client.ts
@@ -7,13 +7,13 @@ import {
   Topic,
   TopicWithMetaData,
 } from '@service-bus-browser/service-bus-api-contracts';
-import { ApiHandler } from './api-handler';
+import { BackendApi } from './backend-api';
 
 export class ServiceBusManagementFrontendClient {
-  constructor(private serviceBusApi: ApiHandler) {}
+  constructor(private backendApi: BackendApi) {}
 
   async listQueues(connectionId: string): Promise<QueueWithMetaData[]> {
-    return (await this.serviceBusApi.serviceBusManagementDoRequest(
+    return (await this.backendApi.serviceBusManagementDoRequest(
       'listQueues',
       {
         connectionId,
@@ -25,35 +25,35 @@ export class ServiceBusManagementFrontendClient {
     connectionId: string,
     queueId: string,
   ): Promise<QueueWithMetaData> {
-    return (await this.serviceBusApi.serviceBusManagementDoRequest('getQueue', {
+    return (await this.backendApi.serviceBusManagementDoRequest('getQueue', {
       connectionId,
       queueId,
     })) as QueueWithMetaData;
   }
 
   async createQueue(connectionId: string, queue: Queue): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest('addQueue', {
+    await this.backendApi.serviceBusManagementDoRequest('addQueue', {
       connectionId,
       queue,
     });
   }
 
   async editQueue(connectionId: string, queue: Queue): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest('editQueue', {
+    await this.backendApi.serviceBusManagementDoRequest('editQueue', {
       connectionId,
       queue,
     });
   }
 
   async removeQueue(connectionId: string, queueId: string): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest('removeQueue', {
+    await this.backendApi.serviceBusManagementDoRequest('removeQueue', {
       connectionId,
       queueId,
     });
   }
 
   async listTopics(connectionId: string): Promise<TopicWithMetaData[]> {
-    return (await this.serviceBusApi.serviceBusManagementDoRequest(
+    return (await this.backendApi.serviceBusManagementDoRequest(
       'listTopics',
       {
         connectionId,
@@ -65,28 +65,28 @@ export class ServiceBusManagementFrontendClient {
     connectionId: string,
     topicId: string,
   ): Promise<TopicWithMetaData> {
-    return (await this.serviceBusApi.serviceBusManagementDoRequest('getTopic', {
+    return (await this.backendApi.serviceBusManagementDoRequest('getTopic', {
       connectionId,
       topicId,
     })) as TopicWithMetaData;
   }
 
   async createTopic(connectionId: string, topic: Topic): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest('createTopic', {
+    await this.backendApi.serviceBusManagementDoRequest('createTopic', {
       connectionId,
       topic,
     });
   }
 
   async editTopic(connectionId: string, topic: Topic): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest('updateTopic', {
+    await this.backendApi.serviceBusManagementDoRequest('updateTopic', {
       connectionId,
       topic,
     });
   }
 
   async removeTopic(connectionId: string, topicId: string): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest('deleteTopic', {
+    await this.backendApi.serviceBusManagementDoRequest('deleteTopic', {
       connectionId,
       topicId,
     });
@@ -96,7 +96,7 @@ export class ServiceBusManagementFrontendClient {
     connectionId: string,
     topicId: string,
   ): Promise<SubscriptionWithMetaData[]> {
-    return (await this.serviceBusApi.serviceBusManagementDoRequest(
+    return (await this.backendApi.serviceBusManagementDoRequest(
       'listSubscriptions',
       {
         connectionId,
@@ -110,7 +110,7 @@ export class ServiceBusManagementFrontendClient {
     topicId: string,
     subscriptionId: string,
   ): Promise<SubscriptionWithMetaData> {
-    return (await this.serviceBusApi.serviceBusManagementDoRequest(
+    return (await this.backendApi.serviceBusManagementDoRequest(
       'getSubscription',
       {
         connectionId,
@@ -125,7 +125,7 @@ export class ServiceBusManagementFrontendClient {
     topicId: string,
     subscription: Subscription,
   ): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest(
+    await this.backendApi.serviceBusManagementDoRequest(
       'createSubscription',
       {
         connectionId,
@@ -140,7 +140,7 @@ export class ServiceBusManagementFrontendClient {
     topicId: string,
     subscription: Subscription,
   ): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest(
+    await this.backendApi.serviceBusManagementDoRequest(
       'updateSubscription',
       {
         connectionId,
@@ -155,7 +155,7 @@ export class ServiceBusManagementFrontendClient {
     topicId: string,
     subscriptionId: string,
   ): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest(
+    await this.backendApi.serviceBusManagementDoRequest(
       'deleteSubscription',
       {
         connectionId,
@@ -171,7 +171,7 @@ export class ServiceBusManagementFrontendClient {
     subscriptionId: string,
     rule: SubscriptionRule,
   ): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest(
+    await this.backendApi.serviceBusManagementDoRequest(
       'addSubscriptionRule',
       {
         connectionId,
@@ -188,7 +188,7 @@ export class ServiceBusManagementFrontendClient {
     subscriptionId: string,
     rule: SubscriptionRule,
   ): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest(
+    await this.backendApi.serviceBusManagementDoRequest(
       'editSubscriptionRule',
       {
         connectionId,
@@ -205,7 +205,7 @@ export class ServiceBusManagementFrontendClient {
     subscriptionId: string,
     ruleName: string,
   ): Promise<void> {
-    await this.serviceBusApi.serviceBusManagementDoRequest(
+    await this.backendApi.serviceBusManagementDoRequest(
       'removeSubscriptionRule',
       {
         connectionId,
