--- libs/api-clients/clients/src/lib/workspaces-api-handler.ts
@@ -1,3 +1,0 @@
-export interface WorkspacesApiHandler {
-  workspacesDoRequest(requestType: string, request: unknown): Promise<unknown>;
-}
